All api and core django files pass linting

+++ backend/music_space/api/signals.py
@@ -1,3 +1,4 @@
+"""signals.py"""
 
 import pathlib
 
@@ -7,9 +8,17 @@ from django.dispatch import receiver
 
 from .models import UserProfile
 
+default_name = 'profile_pics/default.png'
+
 
 @receiver(post_save, sender=User)
-def create_userProfile(sender, instance, created, **kwargs):
+def create_userprofile(
+    sender: type[User],  # noqa: ARG001
+    instance: User,
+    created: bool,  # noqa: FBT001
+    **kwargs: dict,  # noqa: ARG001
+) -> None:
+    """Crea un UserProfile quan es crea un nou User."""
     if created:
         UserProfile.objects.create(
             user=instance,
@@ -19,17 +28,31 @@ def create_userProfile(sender, instance, created, **kwargs):
 
 
 @receiver(pre_save, sender=UserProfile)
-def delete_old_profile_pic(sender, instance, **kwargs):
+def delete_old_profile_pic(
+    sender: type[UserProfile],  # noqa: ARG001
+    instance: UserProfile,
+    **kwargs: dict,  # noqa: ARG001
+) -> None:
+    """Elimina l'arxiu de la imatge de perfil antiga abans de desar una nova."""
     if instance.pk and UserProfile.objects.filter(pk=instance.pk).exists():
         old_file = UserProfile.objects.get(pk=instance.pk).profilePic
         new_file = instance.profilePic
-        if old_file and old_file != new_file and old_file.name != 'profile_pics/default.png':
-            if pathlib.Path(old_file.path).exists():
-                pathlib.Path(old_file.path).unlink()
+        if (
+            old_file and old_file != new_file and old_file.name != default_name
+        ) and pathlib.Path(old_file.path).exists():
+            pathlib.Path(old_file.path).unlink()
 
 
 @receiver(post_delete, sender=UserProfile)
-def delete_profile_pic_on_delete(sender, instance, **kwargs):
-    if instance.profilePic and instance.profilePic.name != 'profile_pics/default.png':
-        if pathlib.Path(instance.profilePic.path).exists():
-            pathlib.Path(instance.profilePic.path).unlink()
+def delete_profile_pic_on_delete(
+    sender: type[UserProfile],  # noqa: ARG001
+    instance: UserProfile,
+    **kwargs: dict,  # noqa: ARG001
+) -> None:
+    """Elimina l'arxiu de la imatge de perfil quan es elimina el UserProfile."""
+    if (
+        instance.profilePic
+        and instance.profilePic.name != default_name
+        and pathlib.Path(instance.profilePic.path).exists()
+    ):
+        pathlib.Path(instance.profilePic.path).unlink()
